requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
Deprecated==1.2.5
enum34==1.1.6
idna==2.8
ipaddress==1.0.22
pycparser==2.19
PyGithub==1.43.7
PyJWT==1.7.1
pyOpenSSL==19.0.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
wrapt==1.11.1
